updateref: Implement ref creation and deletion via `Update()`
When driving git-update-ref(1) with the `--stdin` flag, it supports two
verbs "create" and "delete", where "create" is equivalent to "update"
with the old value being the zero OID, and "delete" is equivalent to
"update" with a zero OID new value.

Let's implement both these commands in terms of `Update()` such that we
deduplicate the logic required to write to the git-update-ref(1)
process.