Don't delete unmerged branches in UpdateRemoteMirror
An old bug was reintroduced in UpdateRemoteMirror while it was being
ported to Go. The Ruby implementation did not delete branches in the
remote that were not present in the local repository if they contained
work not merged in the local repository's default branch. This commit
brings the Go port in line with the previous behavior in the Ruby
implementation and adds a regression test for the bug.

Changelog: fixed