updateref: Rename `Wait()` to `Commit()`
Before we have converted the updateref package to always use
transactional semantics in 154dbc1a2 (updateref: Safeguard against
accidentally committing updates, 2020-12-15), all the `Wait()` function
used to do was to call `Wait()` on git-update-ref(1). This simply closed
stdin and then waited for the command to return.

But since above commit, semantics have changed such that we now always
use an explicit transaction, where `Wait()` is now responsible for
committing the transaction by sending "commit" to the process. In this
context, calling the function `Wait()` doesn't really convey its spirit
anymore: yes, we wait for the command, but what we are really doing is
to commit the transaction.

Rename the function to `Commit()` to better highlight its modern use.