updateref: Unify update of the transactional state
We currently have multiple code paths which update the transactional
state of the git-update-ref(1) command. In theory, updating the state
would involve both sending the command as well as verifying that the
state update has successfully executed, and in such a world it does make
sense to have a single function which handles both such that we don't
have to repeat this logic.

Unfortunately, git-update-ref(1) is currently broken such that it does
not flush its stdout after acknowledging the state change. As a result,
we cannot verify that the update has been successfully processed without
risking a deadlock. But still, this commit unifies code paths via a new
`setState()` function, which for now only writes the state change verb
without checking for success.