repoclean: Background task to act on the repositories
Background task runs periodically and scans gitaly nodes for the
repositories. For each repository it determines if it is known to
praefect or not. Based on this info some action can be performed.
In our first implementation the action would be a warning log
message for the missing repositories (upcoming set of changes).
It uses only healthy gitaly nodes to search for repositories and
gracefully handles any errors that occur during communication
with gitaly nodes.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3719