updateref: Fix quarantine directory being migrated too late
When we have a quarantine directory in place, then we migrate its
objects into the main object database after we have successfully voted
on the ref update via the reference-transaction hook. Most notably, we
do so _after_ we have executed the update hook. This is different from
what Git is doing, which will migrate objects after the pre-receive has
successfully executed. The result is that the update hook runs with a
wrong environment.

This hasn't been much of an issue until now: we don't really use the
update hook ourselves, and custom hooks would have envvars set up such
that it can find quarantined objects. It will become an issue when we
fix our locking: we want to lock the ref update before doing the
"prepared" vote, but do not yet have all objects available in the main
object database.

Fix the issue by aligning our semantics with what Git is doing, which is
to migrate objects directly after the pre-receive hook.

Changelog: fixed