config: Improve error messages of storages validation
The old format of some of the error messages was not too
much friendly to the caller as it included standard
serialization of Go structs. Tests to verify nested repositories
were setup incorrectly and returned error because of not
existing path. It was fixed by using paths to a real directories.
The change also covers more failure cases that may appear as well
as improves error messages.