config: validateShell refactored
When the failure happens because the gitlab-shell.dir is
not set or set incorrectly it is hard to find if it is
it or some other path used in the config file because
it doesn't include any flags to distinguish. The change
adds path to the configuration key, so it is visible in
the error message. Also it makes errors returned by the
method more readable.