config: validateListeners refactored
The validateListeners is not complete as it doesn't
verify if tls_listen_addr is set. So it is not possible
to start gitaly without socket_path or listen_addr set.
We fix it by adding the address to the verification list.
It also removes error prefix as it moved outside of the
validateXXX methods. The corresponding test is refactored
as well to verify error message returned by the message.