git2go: Fix tests for actions overwriting directory
gitaly-git2go(wrapping libgit2) acts as a no-op when adding `IndexEntry`
to index if the `Path` of `IndexEntry` is already a directory in index.
So when creating or moving files that overwrite an existing directory,
`CreateFile` would do nothing and `MoveFile` would just delete the old
file, which is rather weird.

When creating a file with the path of an existing directory on gitlab,
the operation would succeed with an empty commit created. Obviously this
is not what we want.

This commit adjusts corresponding test cases for git2go to ensure we get
error when creating or moving files overwriting existing directories.

Signed-off-by: blanet <moweng.xx@alibaba-inc.com>