command: Create separate tracing span when acquiring spawn tokens
When trying to dig into performance-related issues, one frequently sees
trace spans where commands are seemingly hanging for a long time while
doing nothing. More often than not, the root cause of this is that too
many requests come in and as a result have emptied available spawn
tokens. While this becomes visible as soon as one takes a look at logs,
it is highly misleading when only taking a look at trace spans.

Create a separate span when retrieving spawn tokens to better highlight
what's happening and avoid confusion.