catfile: Get rid of spawn error injection
The catfile BatchCache struct has a simple know for testing purposes
which allows us to simulate errors with spawning the git-cat-file(1)
processes. This know isn't really required nowadays anymore though:
we can just supply a `git.CommandFactory` which raises those errors for
us.

Convert the code to do so and remove the infra for injecting spawn
errors.