catfile: Refactor batch cache tests to use real structs
The `BatchCache` tests are currently working with sentinel values which
are added to and retrieved from the cache, which is simply an empty
`batch` struct. While this is easy to use for tests, it doesn't reflect
real-world use and furthermore keeps us from extending functionality of
the `batch` struct because we always have an uninitialized struct, only.

Fix this by creating a new `mustCreateBatch()` helper function which
creates real `batch` structs for us via `newBatch()`.

This also uncovers a bug in our `addTwice` test, where we used to
compare values retrieved from the cache, but expected the wrong value to
be returned: because all values were empty, comparing them with each
other for equality will always return `true`. This test bug is fixed by
checking for the correct value.