Use repository ID in GetReplicatedGeneration
To prepare for Praefect generating disk paths for replicas, this commit
changes GetReplicatedGeneration to identify the repository by its ID rather
than the virtual storage and relative path. Functionality is the same but
the DowngradeAttemptedError had to be changed as the method no longer has
the information about virtual storage and relative path. This is fine though
as the virtual storage and relative path are available for logging higher up
in the call chain.