catfile: Explicitly close processes
The catfile package is playing quite some tricks with contexts where
they are used both to set up correlation, but also to cancel processes.
While the former one makes sense, going the indirect route via the
context to kill processes is confusing and makes the code hard to reason
about.

Convert the code to explicitly kill processes as required to fix this.