Update SetGeneration to use repository ID
This commit updates SetGeneration to use repository ID to identify
a repository instead of the virtual storage and relative path. Logic
is the same but some tests had minor changes to their setups as they
did not setup the repository IDs correctly. SetGeneration is used to
set a repository's generation after successfully applying a replication
job to it. Using the repository ID instead of the virtual storage and
the relative path guarantee any jobs targeting a previous deleted or
renamed repository at the same virtual storage and relative path do
not affect the repository that is now identified by them.