Register RepositoryExists as a handler instead of interceptor
Praefect intercepts RepositoryExists calls to check whether a repository
should exist from the database rather than relaying to a Gitaly node.
Currently this is handled by a special interceptor that only handles
RepositoryExists calls. There's a better way by registering a handler
directly for the method. This way the interceptor doesn't end up in the
call chain of every RPC and it's easier to add more methods to intercept.
This commit moves the handling from an interceptor to a registered method.