ref: Use exact comparison on page token
The FindLocalBranches accepts a page token in it's request. The caller
can fill this in to get the records from a certain offset. Prior to this
change a lexicographical comparison was used to determine the records
that come *after* the page token. This was done to make sure the page
token will work even if some records were removed in mean time.

This method works fine when the refs are requested in alphabetical
order, but that's not always the case. With this change the comparison
will use an exact match on the page token. When the page token is not
found, it raises an error indicating the page token was not found.

The change is behind a default-disabled feature flag and rolled out
through [1].

1. https://gitlab.com/gitlab-org/gitaly/-/issues/3817