command: Fix in-flight process counter decreasing too late
We keep track of in-flight processes via a simple wait group such that
we can easily assert in our tests that there are no processes around
after tests have concluded. This counter gets decremented in `New()`
after the context has been wrapped up and the command has returned.
While this works alrighth in most cases, some code paths make use of
long-running processes which are decoupled from the original context,
where the new context will never be cancelled. The end result is that
we're needlessly waiting in `testhelper.MustHaveNoChildProcess()` for
two seconds even though all processes have already exited.

Move the call to `wg.Done()` into `wait()`: this function will be
executed either when the context has been cancelled, or explicitly by
the user as soon as the command is expected to terminate. Like this,
decrementing the counter is now decoupled from the context and will thus
provide direct feedback on finished processes, removing the two-second
delay at test exit.