git: Add tests for `CurrentVersion()`
While we test whether we're able to parse version strings correctly, we
do not have any tests which verify that we're able to do so from a
spawned command. This commit fills in this gap by slightly refactoring
`CurrentVersion()` to read the stdout directly from the returned
command. This allows us to easily stub out the CommandFactory with our
own which always echo's a fixed version string.