server: Fix test race caused by concurrent UserCreateTag requests
The `TestAuthBeforeLimit()` functions verifies that the authentication
token's validity is checked before actually executing the RPC logic
itself. This is done by setting up a pre-receive hook that waits for the
tokens' validity duration such that the RPC is guaranteed to take longer
than the token would be valid.

The RPC call that is being executed is `UserCreateTag`, where we try to
create the same tag twice. While we would typically just force-overwrite
the tag and thus don't care about this race, the fact that we're waiting
in the pre-receive hook where refs would still be locked means that both
processes will try to lock the same ref concurrently. We haven't cared
about this in the past given that the updateref package never had
correct locking semantics. But now that we're fixing it to do proper
locking, it will cause the test to fail.

Fix this bug by simply creating two separate tags. We don't care about
the results of the RPC, but only about the artificial delay via the
hook. So changing what is being created is perfectly fine.