proto: Support marking RPC methods as intercepted
Praefect currently supports marking an entire service as being intercepted.
When a service is intercepted, it's not required to have operation type
annotations which are used for proxying as they are not needed. This commit
adds support for marking a single method from a service as intercepted. This
allows for dropping annotations for methods which do not get proxied.