proto: Mark RepositoryExists as an intercepted method
Praefect intercepts RepositoryExists calls and does not pass them through
the proxy to Gitalys. As we can now mark a single method from a service as
intercepted, let's do so for RepositoryExists. This allows us to remove its
operation type annotations which are unused and to clearly indicate it is
intercepted by Praefect.