Intercept RepositoryExists regardless of the election strategy
Praefect only intercepts RepositoryExists if PerRepositoryElector is
enabled. This is so because with SQL elector Praefect could be still
missing database records for repositories. As SQL elector is not usable
anymore, this commit drops this unnecessary condition. This allows us
to drop the operations annotations later as RepositoryExists is always
intercepted and never proxied.

This change also caught RenameRepository test case which was not really
testing a  real scenario due to using a mocked RepositoryStore. The test
case is updated to use a database so the actual behavior gets better
tested.