updateref: Return typed error in case preparing ref update fails
When updating a reference with hooks, then we return a typed `Error`
structure in case the underlying git-update-ref(1) process itself fails.
Until now, this could really only happen when calling `Commit()` given
that we never verified that preparing the ref updates really succeeded.
We're about to fix it though such that we notice early when preparing
the ref update fails.

Prepare the hook updater for this and return a typed `Error` structure
when `Prepare()` fails.