ref: Handle failures when preparing ref transactions fails in DeleteRefs
When updating refs fails in DeleteRefs, then we do not return an error
but instead return a `DeleteRefsResponse` with the embedded `GitError`
field being set to the actual error message. We only do so when calling
the updater's `Commit()` function though, which has worked alright until
now given that `Prepare()` never verified that preparing the transaction
was indeed successful. We're about to fix it though such that we notice
early when preparation of queued reference updates fails.

Prepare the DeleteRefs RPC to handle this correctly by not returning the
error verbosely, but instead by embedding it in the response as we do on
`Commit()`.