coordinator: Split out function to compute metadata for Gitaly peers
When proxying a request to Gitaly, then we need to make sure to pass
relevant metadata to Gitaly. Right now, this only requires us to convert
the incoming metadata into outgoing metadata such that it's correctly
passed on as part of any RPC calls. But we need to extend the logic to
provide better handling for feature flags.

Move the logic into a separate function `streamParametersContext()` to
make it easily extensible.