catfile: Get rid of the separate instrumented batch implementation
We have two implementations of the Batch interface, one which implements
the logic and a second one which wraps it and puts instrumentation on
top. We always return the instrumented version from the cache though, so
it's kind of a moot split and tends to result in confusion when one
reads the code.

Get rid of the separate instrumented batch implementation and move the
relevant pieces directly into the real batch implementation.