featureflag: Expose generation of raw metadata keys
When putting feature flags into the context metadata or extracting them
from it, we need to compute their key based on their name by doing some
smallish transformations. The logic is currently private to the package
though and thus it cannot be reused by other packages.

Implement the `headerKey()` function as receiver of the `FeatureFlag`
struct and expose it such that it's generally available.