ssh: Verify that feature flags get passed correctly on push
Because we do not control execution of gitaly-ssh ourselves but instead
have it executed by git-push(1), we are forced to pass parameters via a
set of serialized environment variables. Next to the payload, which is
the RPC request, we also recognize a GITALY_FEATUREFLAGS variable which
is supposed to contain the list of enabled feature flags.

This mechanism is never really tested in our SSHReceivePack tests. But
with the upcoming change in Praefect which will always inject all known
feature flags into the context, we will by necessity always get these
feature flags serialized into the HooksPayload by Praefect. Given that
we compare the received HooksPayload with our own one and because the
payload will also contain all seen feature flags, this change will cause
the test to break.

Let's prepare for this by explicitly injecting all known feature flags.
This both fixes the test failure, but it also ensures that we can indeed
pass on feature flags as expected to gitaly-ssh.