objectpool: Deprecate `UnlinkRepositoryFromObjectPool()`
The `UnlinkRepositoryFromObjectPool()` function is supposed to unlink a
repository from its object pool. But tall it does is to call `Unlink()`,
which in turn only removes the corresponding remote which pointed to the
repository from the object pool's repository. In fact, it's not even
really doing that given that we never create this remote in the first
place. As can be guessed by this no-op action, this RPC call is never
executed by anything.

Deprecate the RPC such that we can remove it in release 14.4.

Changelog: deprecate