objectpool: Deduplicate setup of test object pool
Creation of object pools requires quite a bunch of different
dependencies, and these are about to be extended to also include the
transaction manager. This commit refactors the objectpool servicee tests
such that this is encapsulated in a simple function `initObjectPool()`,
which sets up required dependencies for us based on the passed config.
This will allow us to change `NewObjectPool()` with less secondary
churn.

Note that ideally, we'd want to have this funciton in our gittest
package. But due to resulting cyclic dependencies, this unfortunately is
not easily possible.