operations: Convert `UserApplyPatch()` to write config transactionally
When applying patches via `UserApplyPatch()`, we first set
"core.splitIndex" before creating a worktree. This is currently done in
a non-transactional way and can thus easily race with concurrent updates
of the same gitconfig.

Convert it to use our new `localrepo.SetConfig()` function, which is
properly aware of transactional semantics.

Changelog: fixed