repository: Convert `MidxRepack()` to write config transactionally
When repacking theh repository with git-multi-pack-index(1), then we
first write the "core.multiPackIndex" key into the gitconfig. This is
currently done in a non-transactional way and can thus easily race with
concurrent updates of the same gitconfig.

Convert it to use our new `localrepo.SetConfig()` function, which is
properly aware of transactional semantics.

Changelog: fixed