localrepo: Provide function to change config entries with transactions
Recently, we have started to up our game when writing to on-disk files
such that they use proper transactional semantics with locking and
voting. The first prototype for this was the `SetFullPath()` RPC call,
which writes a single config entry into the gitconfig file. This
prototype has shown to not cause any issues, so it's now the logical
next step to extend this schema to other callsites which write to the
gitconfig.

As a first preparatory step, this commit extracts the logic from the RPC
call into a generic location, which is the localrepo.