STYLE: Recommend using direct and imperative wording
In the same way as it is recommended for Git development
in the `imperative-mood` section of their
`Documentation/SubmittingPatches.txt` guidelines, we
should recommend using the imperative mood or at least
direct style in commit message subjects.

While at it, let's recommand a direct and 'get to the
point' style in the whole message.