repository: Fix wrong usage of git-commit-graph(1)
Recently, our nightly CI job started to fail with an invalid usage
message when executing git-commit-graph(1). As it turns out, our current
usage of `git commit-graph write --split replace` in one of our tests
didn't do what we'd think it does: `--split replace` is not treated the
same as `--split=replace`, but it is instead treated as a separate
argument. And because of a bug in Git itself, unparsed arguments for the
"write" and "verify" subcommands were completely ignored.

This bug was fixed upstream via 6d209a01f8 (commit-graph: show usage on
"commit-graph [write|verify] garbage", 2021-08-23), which now correctly
causes our test to fail.

Fix the bug by passing `--split=replace` as intended.