repository: Convert `CreateRepositoryFromURL` to be atomic
Now that we have a central helper function which knows to create
repositories atomically, convert `CreateRepositoryFromUR()` to use this
new helper. This changes semantics of the RPC call such that it will
only create repositories if they do not exist yet. Furthermore, no data
will be left behind if the RPC call fails.

Changelog: fixed