test: Close error channel to allow multiple call for cleanup
The cleanup returned by the runPraefectServer reads for an error
from the channel. If it is run multiple times it blocks on the
second run as it doesn't receive data from the channel. As
there is only a single send to the channel and no one sends anything
else into it we close the channel. With that we can call for
cleanup function multiple times without issues.