replication: Graceful stop of the replication processing loop
The loop for processing replication events uses a root context
of the application. Once the context is cancelled in-flight
operation would be cancelled as well (network call, db operation,
etc.). Instead we should allow operation to complete and
stop processing events on the next iteration. It makes execution
of the replication events more stable and deterministic.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2703

Changelog: Fixed