test: Simplify tests waiting for the replication events
The ReplicationEventQueueInterceptor interface is not used in
the production code as well as it's single implementation.
That change simplifies it by removing the interface and making
interceptor a testing entity. Now it tracks input and output
parameters used in methods calls. Now they can be used for
verification in the tests. The test code is simplified as well
because of the new Wait method on the interceptor. The call sites
now simplified based on this new additions.