blob: Convert `GetBlob()` to use object reader
The `GetBlob()` RPC retrieves a single blob by object ID from the
repository's object database. This is implemented via the catfile cache
by first retrieving object info via `Batch.Info()`, and then the object
data is retrieved via another call to `Batch.Blob()` in case the info
says it is indeed a blob. This is inefficient: we spawn two processes
and have two round trips to those processes. The only upside is that we
avoid reading objects which are not a blob, but given that non-blob
objetcs are typically not large-ish it is debatable whether this really
has much of an impact at all.

Refactor the code to use our new cached ObjectReader interface and just
read the object directly. This only requires a single process and avoids
one round trip by just reading the object directly.

Changelog: performance